pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ticket-assist"
version = "1.0.4"
description = "Enhanced Python client for JIRA with better error handling, pagination, and metadata validation."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Avi Zaguri" },]
keywords = ["development", "python", ""]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "build>=1.3.0",
    "certifi>=2026.2.25",
    "colorlog>=6.10.1",
    "custom-python-logger>=2.0.11",
    "flask>=3.1.3",
    "pathlib>=1.0.1",
    "pre-commit>=4.5.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "setuptools>=80.9.0",
    "anthropic>=0.49.0",
    "instructor>=1.7.9",
    "slack-bolt>=1.27.0",
    "slack-sdk>=3.40.1",
    "wheel>=0.45.1",
    "pytest-plugins>=2.0.2",
]

[project.urls]
Homepage = "https://github.com/aviz92/ticket-assist"
Repository = "https://github.com/aviz92/ticket-assist"
Issues = "https://github.com/aviz92/ticket-assist/issues"

[tool.setuptools]
include-package-data = true
zip-safe = false

[tool.setuptools.packages.find]
where = ["."]
include = ["ticket_assist*"]
namespaces = false

[tool.setuptools.package-data]
python_base_toolkit = ["ci/**/*"]
